+++ CHANGELOG.md
@@ -303,3 +303,4 @@
 - 2024-02-22: Add lib, task, follow up, configuration, +2 db
 - 2024-02-23: Metrics discovery, mutate, follow up, qualify, +2
 - 2024-02-23: Pool campaign, geo, README, find email, +6
+- 2024-02-24: Rate leads, geo, campaigns, contacts, +5